Fix ClassCastException in getDevicePixelRatio() method

+++ eyes.appium.java/src/main/java/com/applitools/eyes/appium/EyesAppiumDriver.java
@@ -64,7 +64,7 @@ public class EyesAppiumDriver extends EyesWebDriver {
     }
 
     public double getDevicePixelRatio () {
-        return ((Long) getCachedSessionDetails().get("pixelRatio")).doubleValue();
+        return (Double) getCachedSessionDetails().get("pixelRatio");
     }
 
     /**
